Raking sprint and errors fix

+++ src/AppBundle/Controller/HomeController.php
@@ -10,13 +10,17 @@ use AppBundle\Entity\Post;
 use AppBundle\Entity\PostLike;
 use AppBundle\Entity\Following;
 use AppBundle\Entity\User;
+use AppBundle\Entity\GameGroupVote;
 use AppBundle\Form\PostType;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
+
 class HomeController extends Controller
 {
     /**
+     * Esta función generará la información para mostrar en la vista principal de la aplicación una vez
+     * el usuario se ha autenticado
      * @Route("/home", name="homepage")
      */
     public function indexAction(Request $request)
@@ -26,20 +30,24 @@ class HomeController extends Controller
         $repository_post = $this->getDoctrine()->getRepository(Post::class);
         $repository_following = $this->getDoctrine()->getRepository(Following::class);
         $repository_likes = $this->getDoctrine()->getRepository(PostLike::class);
+        $repository_votes = $this->getDoctrine()->getRepository(GameGroupVote::class);
         
         $logged_user = $this->container->get('security.token_storage')->getToken()->getUser();
 
+        // Obtenemos las publicaciones realizadas por el usuario logueado y por los usuarios que
+        // el mismo está siguiendo
         $query_post = $repository_post->findPostUserAndFollowing($logged_user);
 
-        $query_numpost_user = $repository_post->getNumPostsUser($logged_user);
-        $query_numfollowing_user = $repository_following->getNumFollowingsUser($logged_user);
-        $query_numfollowers_user = $repository_following->getNumFollowersUser($logged_user);
+        $query_numpost_user = $repository_post->getNumPostsUser($logged_user); // Número de publicaciones del usuario
+        $query_numfollowing_user = $repository_following->getNumFollowingsUser($logged_user); // Número de usuarios que sigue
+        $query_numfollowers_user = $repository_following->getNumFollowersUser($logged_user); // Número de seguidores que posee
+        $query_puntuation = $repository_votes->getAverageVoteUser($logged_user->getId()); // Puntuación media que posee de los grupos
 
         foreach($query_post as $post) {
 
-            $datetime = $this->timeago($post->getDatetime());
-            $likes = $repository_likes->getNumLikesPost($post);
-            $has_like = $repository_likes->getUserHasLikePost($post, $logged_user);
+            $datetime = $this->timeago($post->getDatetime()); // Tiempo que ha pasado desde que se realizó la publicación
+            $likes = $repository_likes->getNumLikesPost($post); // Likes de la publicación
+            $has_like = $repository_likes->getUserHasLikePost($post, $logged_user); // Con eso sabemos si el usuario le ha dado me gusta a la publicación
             $posts[] = array(
                 'id' => $post->getId(),
                 'user' => $post->getUser(),
@@ -51,6 +59,7 @@ class HomeController extends Controller
             );
         }
 
+        // Creamos el formulario que le permitirá al usuario añadir nuevas publicaciones
         $post = new Post();
         $form = $this->createForm(PostType::class, $post);
 
@@ -62,17 +71,18 @@ class HomeController extends Controller
         if($form->isSubmitted()) {
             
             if (count($errors) > 0) {
-                return $this->render('@App/index.html.twig', array(
+                return $this->render('@App/home/index.html.twig', array(
                     'errors' => $errors, 'form' => $form->createView()
                 ));
             } else {
 
-                // upload file
+                // Imagen añadida a la publicación
                 $file = $form["image"]->getData();
                 $ext = $file->guessExtension();
                 $file_name = time().".".$ext;
-                $file->move("uploads",$file_name);
+                $file->move("uploads/posts/",$file_name);
 
+                // Almacenamos la información de la publicación en la base de datos
                 $user = $this->getUser();
                 $post->setUser($user);
                 $post->setDatetime(new \DateTime('now'));
@@ -80,20 +90,27 @@ class HomeController extends Controller
                 $entityManager = $this->getDoctrine()->getManager();
                 $entityManager->persist($post);
                 $entityManager->flush();
+
                 $this->get('session')->getFlashBag()->add('notice', 'Su publicación se ha creado correctamente');
                 return $this->redirectToRoute('homepage');
             }
         }
 
-        if(isset($posts))
-            return $this->render('@App/index.html.twig', array('posts' => $posts, 'numposts' => $query_numpost_user,
-                'numfollowing' => $query_numfollowing_user, 'numfollowers' => $query_numfollowers_user,
+        return $this->render('@App/home/index.html.twig', 
+            array(
+                'posts' => isset($posts)? $posts:null,
+                'numposts' => isset($query_numpost_user) ? $query_numpost_user:0,
+                'numfollowing' => isset($query_numfollowing_user) ? $query_numfollowing_user:0,
+                'numfollowers' => isset($query_numfollowers_user) ? $query_numfollowers_user:0,
+                'puntuation' => isset($query_puntuation) ? $query_puntuation:0,
                 'form' => $form->createView()));
-        else
-            return $this->render('@App/index.html.twig', array("form" => $form->createView()));
     }
 
-    public function timeago($datetime) {
+    /**
+     * Función que nos permite calcular cuanto tiempo ha pasado desde que se realizó una publicación
+     * para así mostrarlo en la misma
+     */
+    private function timeago($datetime) {
         date_default_timezone_set('Europe/Madrid');
         $timestamp = strtotime($datetime->format('d-m-Y H:i:s'));
 
@@ -101,10 +118,11 @@ class HomeController extends Controller
         $length = array("60","60","24","30","12","10");
 
         $currentTime = time();
+
         if($currentTime >= $timestamp) {
             $diff     = time()- $timestamp;
             for($i = 0; $diff >= $length[$i] && $i < count($length)-1; $i++) {
-            $diff = $diff / $length[$i];
+                $diff = $diff / $length[$i];
             }
 
             $diff = round($diff);
@@ -113,29 +131,40 @@ class HomeController extends Controller
     }
 
     /**
+     * Función utilizada para dar me gusta a una publicación
      * @Route("/home/likePost/{id_post}", name="like_post", options={"expose"=true})
      */
     public function likePost($id_post) {
- 
+        date_default_timezone_set('Europe/Madrid');
         $repository_post = $this->getDoctrine()->getRepository(Post::class);
         $repository_postlike = $this->getDoctrine()->getRepository(PostLike::class);
 
         $post = $repository_post->findOneBy(array('id' => $id_post));
-        $user = $this->getUser();
 
-        $postlike = new PostLike();
+        if($post) {
+            $postlike = $repository_postlike->findOneBy(array('post' => $id_post, 'user' => $this->getUser()->getId()));
+            
+            if(!$postlike) {
+                $user = $this->getUser();
 
-        $postlike->setUser($user);
-        $postlike->setPost($post);
+                $postlike = new PostLike();
 
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->persist($postlike);
-        $entityManager->flush();
+                $postlike->setUser($user);
+                $postlike->setPost($post);
+                $postlike->setDatetime(new \Datetime('now'));
 
-        return new Response();
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($postlike);
+                $entityManager->flush();
+            }
+                return new Response();
+        }
+        else
+            return $this->render('@App/error_page.html.twig');
     }
 
     /**
+     * Función para quitar un me gusta de una publicación
      * @Route("/home/unlikePost/{id_post}", name="unlike_post", options={"expose"=true})
      */
     public function unlikePost($id_post) {
@@ -148,14 +177,19 @@ class HomeController extends Controller
 
         $postlike = $repository_postlike->findOneBy(array('post' => $post, 'user' => $user));
 
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($postlike);
-        $entityManager->flush();
-        
-        return new Response();
+        if($postlike) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($postlike);
+            $entityManager->flush();
+            
+            return new Response();
+        }
+        else
+            return $this->render('@App/error_page.html.twig');
     }
 
     /**
+     * Función que nos permite generar el nombre de un archivo de manera única
      * @return string
      */
     private function generateUniqueFileName()
